NIR-tree redux
* Large commit where both the NIR-Tree and R+-Tree were rewritten. The NIR-Tree is now much closer to the R+-Tree and the R+-Tree's implementation now uses a different bounding box storage paradigm. Attendant tests and investigation utilities were added or changed.

* New experiment script for lots of experiments.

* A new visualization class that can handle much larger numbers of rectangles and points than Pencil. It does this by directly creating (large) bmp files.

* Fully parameterized the random points benchmark and made it extensible. Removed all dependancies on, and references to, the pencil printer since it is obsoleted by the BMP printer.

* Generalized explicit two dimensional code to work with n dimensions.

* Added a metric to decide between split axes.

* Added polygons back in with a new approach to separating them on node split.

* Added an optimization to not double compute expansion areas.

* Optimized inserts, added new split metric, and a validation function.

* Added more detailed debugging levels. Fixed a typo in function name. Added geometry unit tests. Fixed some bugs when choosing node.

* Changed all float types to double types. Added more point operators. Added unit tests to fully cover Point and Rectangle types. Integrated Biological and California pre-generated datasets into benchmark. Unified statistics gathering and reporting across R/R+/NIR+ trees. Fixed NIR node splitting.

* Aligned R+ tree splitting with algorithm described in paper.

* Integrated rectangular queries for biological dataset.

* Added 5D forestry dataset and fixed bugs in R+ tree causing nodes to exceed their maximum allowed branch factor. Removed a lot of debugging code because I discovered gdb.

+++ src/include/util/graph.h
@@ -12,14 +12,13 @@
 #include <unordered_set>
 #include <util/geometry.h>
 #include <util/debug.h>
-#include <nirtree/pencilPrinter.h>
 
 class Graph
 {
 	private:
 		struct SchedulePoint
 		{
-			float coord;
+			double coord;
 			unsigned index;
 			bool start;
 		};
